requirements.txt
requests
python-dotenv
pandas
streamlit